added JSON reading ability, including serializable objects
git-svn-id: svn://svn.open-ils.org/ILS/trunk@7235 dcc99617-32d9-48b4-a31d-7c20da2025e4

+++ OpenSRF/src/java/org/opensrf/net/xmpp/XMPPReader.java
@@ -130,7 +130,7 @@ public class XMPPReader implements Runnable {
      * timeout is negative, waits potentially forever.
      * @return The number of milliseconds in wait
      */
-    public synchronized long waitCoreEvent(int timeout) {
+    public synchronized long waitCoreEvent(long timeout) {
 
         if(msgQueue.peek() != null || timeout == 0) return 0;
